+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Send `Offer(Agreement)` activity to start payment process.
 - Implement handler for `Offer(Agreement)` activities.
 - Implement handler for `Accept(Offer)` activities.
+- Track status of remote subscriptions.
 
 ### Changed